base auth added, database schema added.

+++ MonoliteUnicorn/EndPoints/Balances/DeleteTransactionEndPoint.cs
@@ -10,9 +10,10 @@ public class DeleteTransactionEndPoint : ICarterModule
     public void AddRoutes(IEndpointRouteBuilder app)
     {
         app.MapDelete("/balances/transaction/{id}",
-                async (ISender sender, HttpContext context, string id, CancellationToken token) =>
+                async (ISender sender, ClaimsPrincipal claims, string id, CancellationToken token) =>
                 {
-                    var userId = context.User.FindFirst(ClaimTypes.NameIdentifier)?.Value ?? "";
+                    if (!Guid.TryParse(claims.FindFirstValue(ClaimTypes.NameIdentifier), out var userId)) 
+                        return Results.Unauthorized();
                     var command = new DeleteTransactionCommand(id, userId);
                     await sender.Send(command, token);
                     return Results.Ok();
